// web/src/api/generated/models/ApiV1Vector.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api/v1/user.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ApiV1VectorIDs } from './ApiV1VectorIDs';
import {
    ApiV1VectorIDsFromJSON,
    ApiV1VectorIDsFromJSONTyped,
    ApiV1VectorIDsToJSON,
    ApiV1VectorIDsToJSONTyped,
} from './ApiV1VectorIDs';

/**
 * Vector is an element used in the Differential Evoluition algorithm.
 * @export
 * @interface ApiV1Vector
 */
export interface ApiV1Vector {
    /**
     * 
     * @type {ApiV1VectorIDs}
     * @memberof ApiV1Vector
     */
    ids?: ApiV1VectorIDs;
    /**
     * 
     * @type {Array<number>}
     * @memberof ApiV1Vector
     */
    elements?: Array<number>;
    /**
     * 
     * @type {Array<number>}
     * @memberof ApiV1Vector
     */
    objectives?: Array<number>;
    /**
     * 
     * @type {number}
     * @memberof ApiV1Vector
     */
    crowdingDistance?: number;
}

/**
 * Check if a given object implements the ApiV1Vector interface.
 */
export function instanceOfApiV1Vector(value: object): value is ApiV1Vector {
    return true;
}

export function ApiV1VectorFromJSON(json: any): ApiV1Vector {
    return ApiV1VectorFromJSONTyped(json, false);
}

export function ApiV1VectorFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiV1Vector {
    if (json == null) {
        return json;
    }
    return {
        
        'ids': json['ids'] == null ? undefined : ApiV1VectorIDsFromJSON(json['ids']),
        'elements': json['elements'] == null ? undefined : json['elements'],
        'objectives': json['objectives'] == null ? undefined : json['objectives'],
        'crowdingDistance': json['crowdingDistance'] == null ? undefined : json['crowdingDistance'],
    };
}

export function ApiV1VectorToJSON(json: any): ApiV1Vector {
    return ApiV1VectorToJSONTyped(json, false);
}

export function ApiV1VectorToJSONTyped(value?: ApiV1Vector | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'ids': ApiV1VectorIDsToJSON(value['ids']),
        'elements': value['elements'],
        'objectives': value['objectives'],
        'crowdingDistance': value['crowdingDistance'],
    };
}